+++ src/components/map.js
@@ -2,8 +2,8 @@ import React, {Component} from 'react'
 import GoogleMapReact from 'google-map-react'; //see https://github.com/google-map-react/google-map-react/blob/master/API.md
 
 import Marker from './marker'; // Marker object used to track locations
-import NewMarker from './newMarker';
-import SearchBox from './searchBox';
+import NewMarker from './newMarker'; // Separate Marker used for place lookup or new marker submission
+import SearchBox from './searchBox'; // Window attached to map for adding markers, place lookup, and place list
 
 import "./css/map.css"
 import MapStyle from "./css/mapStyle"
@@ -29,6 +29,7 @@ class GoogleMap extends Component {
         this.showPlaceSearch = this.showPlaceSearch.bind(this);
         
         this.state = {
+            //list of existing markers on the map
             markers: [
                 {lat: 34.08421909476845, lng: -118.07298836096781, name:"In-N-Out", description:"Cheap Meals", show: false, id: 0}, // in-n-out
                 {lat: 34.07993604059942, lng: -118.08234390563354, name:"Bay Island", description:"Good Chinese Food", show: false, id: 1}, // bay island
@@ -36,7 +37,8 @@ class GoogleMap extends Component {
                 {lat: 34.10543567839181, lng: -118.07300981856079, name:"Green Zone", description:"Bougie Organic Food", show: false, id: 3}, // green zone
                 {lat: 34.0897531, lng: -118.0529848, name:"Popeyes", description:"Chicken. Need I say more?", show: false, id: 4}, // popeyes
             ],
-            placeList:[],
+
+            placeList:[], //list of markers to be shown on the map *references markers above and is not a copy
 
             //map bools
             addMarkerMode: false,
@@ -111,6 +113,8 @@ class GoogleMap extends Component {
         for(var i = 0; i < markers.length; i++){
             if(i === index){ //index matches
                 markers[i].show = !markers[i].show;
+                console.log(markers);
+                console.log(this.state.placeList);
             } else { //close other markers
                 markers[i].show = false;
             }
@@ -168,20 +172,19 @@ class GoogleMap extends Component {
             ceil.lng = floor.lng;
             floor.lng = temp;
         }
-        console.log(floor);
-        console.log(ceil);
+        // console.log(floor);
+        // console.log(ceil);
 
         var markers = this.state.markers;
         var placeList = [];
-        for(var i = 0; i < markers.length; i++){
+        for(var i = 0; i < markers.length; i++){ //check if the marker should be added to the list
             if(markers[i].lat > floor.lat && markers[i].lat < ceil.lat){
-                if(markers[i].lng > floor.lng && markers[i].lng < ceil.lng){
-                    console.log("Should add");
-                    placeList = placeList.concat(markers[i]);
+                if(markers[i].lng > floor.lng && markers[i].lng < ceil.lng){ 
+                    placeList = placeList.concat(markers[i]); //add marker to visible list if passes both checks
                 }
             }
         }
-        console.log(placeList);
+        //console.log(placeList);
         this.setState({placeList: placeList});
     }
 
@@ -252,6 +255,7 @@ class GoogleMap extends Component {
                     newMarker={this.state.newMarker}
                     showPlaceSearch={this.showPlaceSearch}
                     placeList={this.state.placeList}
+                    openMarker={this.openMarker}
                 />
             </div>
         );