+++ src/components/placeList.js
@@ -3,6 +3,20 @@ import React, {Component} from 'react';
 import "./css/placeList.css";
 
 class PlaceList extends Component{
+    constructor(props){
+        super(props);
+
+        this.state = {
+            place:""
+        }
+
+        this.chooseMarker = this.chooseMarker.bind(this);
+    }
+
+    chooseMarker(e){
+        this.props.openMarker(e.target.id);
+    }
+
     render(){
         var placeList = this.props.placeList
         return(
@@ -10,9 +24,14 @@ class PlaceList extends Component{
                 <span className="PlaceListHeader">
                     Places You're Seeing
                 </span>
-                <hr/>
-                {placeList.map((marker, index) => (
-                    <div key={index}>
+                <hr className="PlaceListDivide"/>
+                {placeList.map((marker) => (
+                    <div 
+                        className={`PlaceEntry ${marker.show ? "active": ""}`}
+                        onClick={this.chooseMarker}
+                        key={marker.id} //required to prevent map errors
+                        id={marker.id} //passes id for marker index
+                    >
                         {marker.name}
                     </div>
                     ))}